Moving InputBlock to inputs
